+++ shoop_tests/front/test_checkout.py
@@ -5,11 +5,15 @@
 #
 # This source code is licensed under the AGPLv3 license found in the
 # LICENSE file in the root directory of this source tree.
+import pytest
 from django.test import override_settings
 
-from shoop.front.checkout.addresses import AddressForm
+from shoop.core.models import get_person_contact
+from shoop.front.checkout.addresses import AddressesPhase, AddressForm
 from shoop.front.checkout.single_page import \
     AddressForm as SinglePageAddressForm
+from shoop.testing.factories import get_default_shop
+from shoop.testing.utils import apply_request_middleware
 
 
 def test_checkout_addresses_has_no_default_country():
@@ -51,3 +55,20 @@ def test_required_address_fields():
         assert form.fields["email"].required == True
         assert form.fields["email"].help_text == "Enter email"
         assert form.fields["phone"].help_text == "Enter phone"
+
+
+@pytest.mark.django_db
+def test_address_phase_authorized_user(rf, admin_user):
+    request = apply_request_middleware(rf.get("/"), shop=get_default_shop(), customer=get_person_contact(admin_user))
+    view_func = AddressesPhase.as_view()
+    resp = view_func(request)
+    assert 'company' not in resp.context_data['form'].form_defs
+
+
+@pytest.mark.django_db
+def test_address_phase_anonymous_user(rf):
+    request = apply_request_middleware(rf.get("/"), shop=get_default_shop())
+    view_func = AddressesPhase.as_view()
+    resp = view_func(request)
+    assert 'company' in resp.context_data['form'].form_defs
+